Make MultiResolutionRenderer generic
